requirements.txt
discord.py>=2.3.2
qdrant-client>=1.7.0
ollama>=0.1.4
